convert to BigDecimal

+++ src/main/java/org/andork/compass/plot/LocationCommand.java
@@ -1,22 +1,24 @@
 package org.andork.compass.plot;
 
+import java.math.BigDecimal;
+
 public interface LocationCommand extends CompassPlotCommand {
 	/**
 	 * @return distance from station to floor, or NaN if missing or passage
 	 */
-	public double getDown();
+	public BigDecimal getDown();
 
 	/**
 	 * @return distance from station to left wall or NaN if missing or passage
 	 */
-	public double getLeft();
+	public BigDecimal getLeft();
 
 	public Location getLocation();
 
 	/**
 	 * @return distance from station to right wall or NaN if missing or passage
 	 */
-	public double getRight();
+	public BigDecimal getRight();
 
 	/**
 	 * @return the name of the station
@@ -26,15 +28,15 @@ public interface LocationCommand extends CompassPlotCommand {
 	/**
 	 * @return distance from station to ceiling or NaN if missing or passage
 	 */
-	public double getUp();
+	public BigDecimal getUp();
 
-	public void setDown(double down);
+	public void setDown(BigDecimal down);
 
-	public void setLeft(double left);
+	public void setLeft(BigDecimal left);
 
-	public void setRight(double right);
+	public void setRight(BigDecimal right);
 
 	public void setStationName(String stationName);
 
-	public void setUp(double up);
+	public void setUp(BigDecimal up);
 }
